Extract protein database actions from MS2Controller
